backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.25.2
redis==5.0.1
celery==5.3.4
pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
azure-cognitiveservices-vision-computervision==0.9.0
google-cloud-vision==3.4.5
boto3==1.34.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0